feat(trace-eap-waterfall): Addressing PR suggestions
